WIP 2 Variable Entities w Tests
