+++ javatests/arcs/core/entity/DummyVariableEntity.kt
@@ -2,29 +2,29 @@ package arcs.core.entity
 
 import arcs.core.data.*
 
-class DummyVariableEntity : VariableEntityBase(ENTITY_CLASS_NAME, SCHEMA), Storable {
-    var bool: Boolean? by SingletonProperty()
-    var num: Double? by SingletonProperty()
-    var text: String? by SingletonProperty()
-    var ref: Reference<DummyVariableEntity>? by SingletonProperty()
-    var bools: Set<Boolean> by CollectionProperty()
-    var nums: Set<Double> by CollectionProperty()
-    var texts: Set<String> by CollectionProperty()
-    var refs: Set<Reference<DummyVariableEntity>> by CollectionProperty()
+class DummyVariableEntity : VariableEntityBase(ENTITY_CLASS_NAME, SCHEMA), Storable, Dummy<DummyVariableEntity> {
+    override var bool: Boolean? by SingletonProperty()
+    override var num: Double? by SingletonProperty()
+    override var text: String? by SingletonProperty()
+    override var ref: Reference<DummyVariableEntity>? by SingletonProperty()
+    override var bools: Set<Boolean> by CollectionProperty()
+    override var nums: Set<Double> by CollectionProperty()
+    override var texts: Set<String> by CollectionProperty()
+    override var refs: Set<Reference<DummyVariableEntity>> by CollectionProperty()
 
     private val nestedEntitySpecs = mapOf(SCHEMA_HASH to DummyVariableEntity)
 
-    fun getSingletonValueForTest(field: String) = super.getSingletonValue(field)
+    override fun getSingletonValueForTest(field: String) = super.getSingletonValue(field)
 
-    fun getCollectionValueForTest(field: String) = super.getCollectionValue(field)
+    override fun getCollectionValueForTest(field: String) = super.getCollectionValue(field)
 
-    fun setSingletonValueForTest(field: String, value: Any?) =
+    override fun setSingletonValueForTest(field: String, value: Any?) =
         super.setSingletonValue(field, value)
 
-    fun setCollectionValueForTest(field: String, values: Set<Any>) =
+    override fun setCollectionValueForTest(field: String, values: Set<Any>) =
         super.setCollectionValue(field, values)
 
-    fun deserializeForTest(rawEntity: RawEntity) = super.deserialize(rawEntity, nestedEntitySpecs)
+    override fun deserializeForTest(rawEntity: RawEntity) = super.deserialize(rawEntity, nestedEntitySpecs)
 
     companion object : EntitySpec<DummyVariableEntity> {
         override fun deserialize(data: RawEntity) =
