+++ javatests/arcs/core/entity/VariableEntityBaseTest.kt
@@ -1,6 +1,8 @@
 package arcs.core.entity
 
-import org.junit.After
+import arcs.core.crdt.VersionMap
+import arcs.core.storage.testutil.DummyStorageKey
+import com.google.common.truth.Truth.assertThat
 import org.junit.Before
 import org.junit.Test
 import org.junit.runner.RunWith
@@ -9,24 +11,38 @@ import org.junit.runners.JUnit4
 
 @RunWith(JUnit4::class)
 @Suppress("UNCHECKED_CAST")
-class VariableEntityBaseTest {
-
-    private lateinit var entity: DummyVariableEntity
+class VariableEntityBaseTest  : EntityTestBase<DummyVariableEntity>() {
 
     @Before
-    fun setUp() {
+    override fun setUp() {
         SchemaRegistry.register(DummyVariableEntity.SCHEMA)
         entity = DummyVariableEntity()
     }
 
-    @After
-    fun tearDown() {
-        SchemaRegistry.clearForTest()
+    @Test
+    override fun serializeRoundTrip() {
+        with (entity) {
+            text = "abc"
+            num = 12.0
+            bool = true
+            ref = createReference("foo")
+            texts = setOf("aa", "bb")
+            nums = setOf(1.0, 2.0)
+            bools = setOf(true, false)
+            refs = setOf(createReference("ref1"), createReference("ref2"))
+        }
+
+        val rawEntity = entity.serialize()
+        val deserialized = DummyVariableEntity()
+        deserialized.deserializeForTest(rawEntity)
+
+        assertThat(deserialized).isEqualTo(entity)
+        assertThat(deserialized.serialize()).isEqualTo(rawEntity)
     }
 
-    @Test
-    fun serializationRoundTrip() {
-        // TODO(alxr): Implement
-    }
+    override fun createReference(id: String) = Reference(
+        DummyVariableEntity,
+        arcs.core.storage.Reference(id, DummyStorageKey(id), VersionMap("id" to 1))
+    )
 
 }
